@@ -1,21 +1,17 @@
 {
     "private": true,
     "scripts": {
-        "dev": "npm run development",
-        "development": "mix",
-        "watch": "mix watch",
-        "watch-poll": "mix watch -- --watch-options-poll=1000",
-        "hot": "mix watch --hot",
-        "prod": "npm run production",
-        "production": "mix --production"
+        "dev": "vite",
+        "build": "vite build"
     },
     "devDependencies": {
         "@tailwindcss/ui": "^0.3",
+        "@vitejs/plugin-vue": "^4.0.0",
         "autoprefixer": "^9.6",
         "axios": "^0.21",
         "bootstrap": "^4.0.0",
         "jquery": "^3.2",
-        "laravel-mix": "^6.0.18",
+        "laravel-vite-plugin": "^0.7.2",
         "lodash": "^4.17.19",
         "popper.js": "^1.12",
         "postcss": "^8.1.14",
@@ -25,14 +21,14 @@
         "sass": "^1.20.1",
         "sass-loader": "^8.0.0",
         "tailwindcss": "^1.8",
-        "vue": "^2.5.17",
+        "vite": "^4.0.2",
+        "vue": "^3.2.25",
         "vue-loader": "^15.9.5",
         "vue-template-compiler": "^2.6.10"
     },
     "dependencies": {
         "laravel-echo": "^1.11.3",
         "pusher-js": "^7.0.3",
-        "vue": "^2.6.12",
-        "vue-chat-scroll": "^1.4.0"
+        "vue": "^3.2.25"
     }
 }
