@@ -4,12 +4,12 @@
  * building robust, powerful web applications using Vue and Laravel.
  */
 
-require('./bootstrap');
+import './bootstrap';
 
 import Vue from 'vue'
 window.Vue = require('vue').default;
 
-const bus = new Vue();      
+const bus = new Vue();
 
 export default bus;
 
