@@ -9,17 +9,14 @@
 
     <title>{{ config('app.name', 'Laravel') }}</title>
 
-    <!-- Scripts -->
-    <script src="{{ asset('js/app.js') }}" defer></script>
-    
     <!-- Styles -->
-    <link href="{{ mix('css/app.css') }}" rel="stylesheet">
+    @vite(['resources/css/app.css', 'resources/js/app.js'])
 </head>
 <body class="text-gray-600 bg-gray-100 h-screen antialiased leading-none font-sans">
     <div id="app">
         <form id="logout-form" action="{{route('logout')}}" method="POST" class="hidden">@csrf</form>
         <header>
-         <navigation-component :user="{{auth()->user()}}"/>   
+         <navigation-component :user="{{auth()->user()}}"/>
         </header>
         @yield('content')
     </div>
